+++ federation/tests/django/utils.py
@@ -1,22 +1,20 @@
 from federation.entities.base import Profile
 
 
-def get_object_function(object_id):
+def dummy_profile():
     return Profile(
         url=f"https://example.com/profile/1234/",
         atom_url=f"https://example.com/profile/1234/atom.xml",
-        id=f"https://example.com/profile/1234/",
+        id=f"https://example.com/p/1234/",
         handle="foobar@example.com",
         guid="1234",
         name="Bob Bobértson",
     )
 
 
+def get_object_function(object_id):
+    return dummy_profile()
+
+
 def get_profile(handle=None, request=None):
-    return Profile(
-        url=f"https://example.com/profile/1234/",
-        atom_url=f"https://example.com/profile/1234/atom.xml",
-        id=f"https://example.com/profile/1234/",
-        handle="foobar@example.com",
-        guid="1234",
-    )
+    return dummy_profile()